// src/types/raw-api-data-type/track/get-track-data-type.ts
export interface GetTrack {
  album?: Album
  artists?: Artist[]
  disc_number?: number
  duration_ms?: number
  explicit?: boolean
  external_ids?: ExternalIDS
  external_urls: ExternalUrls
  href: string
  id: string
  is_local?: boolean
  is_playable?: boolean
  name: string
  popularity: number
  preview_url?: string
  track_number?: number
  type: string
  uri: string
  available_markets?: string[]
  restrictions?: object
}

export interface Album {
  album_type: string
  artists: Artist[]
  external_urls: ExternalUrls
  href: string
  id: string
  images: Image[]
  is_playable?: boolean
  name: string
  release_date: string
  release_date_precision: string
  total_tracks: number
  type: string
  uri: string
  available_markets?: string[]
  restrictions?: object
}

export interface Artist {
  external_urls: ExternalUrls
  href: string
  id: string
  name: string
  type: string
  uri: string
}

export interface ExternalUrls {
  spotify: string
}

export interface Image {
  height: number
  url: string
  width: number
}

export interface ExternalIDS {
  isrc: string
}

export type { GetTrack as TrackItem }